requirements.txt
streamlit
scikit-learn
xgboost
pandas
numpy
seaborn
matplotlib
joblib
pyngrok
google-generativeai
